// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  analytics: typeof analytics;
  api: typeof api_;
  embeddings: typeof embeddings;
  http: typeof http;
  messages: typeof messages;
  rag: typeof rag;
  search: typeof search;
  sessions: typeof sessions;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;